+++ assets/js/main.js
@@ -14,13 +14,4 @@ $(document).ready(function () {
     $('#dueDate').datepicker({
         uiLibrary: 'bootstrap4'
     });
-
-    $('#profile').on('click', function(e) {
-        e.preventDefault();
-        $('#profile-side-bar').toggleClass('active');
-    });
-
-    $('.side-bar-toggle').on('click', function() {
-        $('#profile-side-bar').removeClass('active');
-    });
 });
