+++ portal/urls.py
@@ -9,7 +9,7 @@ from django.views.generic.base import TemplateView
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/', include('invoices.endpoints')),
-    path('api/', include('users.endpoints')),
+    path('api/', include('users.urls')),
 ]+ static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
 
 urlpatterns += [
