+++ users/serializers.py
@@ -16,10 +16,3 @@ class UserSerializer(serializers.ModelSerializer):
             'image',
         )
     read_only_fields = ('image', 'last_name')
-
-    def update(self, instance, validated_data):
-        instance.email = validated_data.get('email', instance.email)
-        instance.first_name = validated_data.get('first_name', instance.first_name)
-        instance.last_name = validated_data.get('last_name', instance.last_name)
-        instance.save()
-        return instance
